created BaseLayout, HomePage and NotFoundPage components. Installed react-router-dom
