+++ src/components/main/main.js
@@ -1,16 +1,9 @@
 import styles from './style.module.css';
-import BurgerConstructor from '../burger-constructor/burger-constructor';
-import BurgerIngredients from '../burger-ingredients/burger-ingredients';
-import { DndProvider } from 'react-dnd';
-import { HTML5Backend } from 'react-dnd-html5-backend';
 
-function Main() {
+function Main({ children }) {
     return (
         <main className={styles.main}>
-            <DndProvider backend={HTML5Backend}>
-                <BurgerIngredients />
-                <BurgerConstructor />
-            </DndProvider>
+            { children }
         </main>
     );
 }
