+++ src/components/main/style.module.css
@@ -1,6 +1,3 @@
 .main {
-    width: 1240px;
-    margin: 0 auto 40px;
-    display: flex;
-    justify-content: space-between;
+    margin-bottom: 40px;
 }
