+++ src/pages/not-found/not-found.js
@@ -1,0 +1,15 @@
+import styles from './style.module.css';
+
+function NotFoundPage() {
+    return (
+        <div className={styles.wrapper}>
+            <div>
+                <p className="text text_type_main-large">
+                    404. Страница не найдена.
+                </p>
+            </div>
+        </div>
+    )
+}
+
+export default NotFoundPage;
