+++ src/pages/not-found/style.module.css
@@ -1,0 +1,6 @@
+.wrapper {
+    min-height: 400px; 
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
